ostatnie poprawki

+++ main/views.py
@@ -1,29 +1,7 @@
 from django.shortcuts import render, redirect
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponseRedirect
 from .models import ToDoList, Item
 from .forms import CreateNewList
-from datetime import datetime
-# Create your views here. tworzenie widoków
-from .serializers import ItemSerializer
-from rest_framework.generics import RetrieveAPIView
-from rest_framework.permissions import DjangoModelPermissions
-
-# class ItemView(RetrieveAPIView):
-#         http_method_names = ["get","patch"]
-#         permission_classes = [DjangoModelPermissions]
-#         serializer_class = ItemSerializer
-
-#         lookup_field = "id"
-                
-#         def get_queryset(self):
-#                 print("tutaj")
-#                 return Item.objects.all()
-
-def remove_item(item_id):
-        # print(f"ls id:{ls_id}, item id:{item_id}")
-        Item.objects.filter(id=item_id).delete()
-        # return redirect(to=f"/{ls_id}")
-
 
 def new_item(response,id):
         ls = ToDoList.objects.get(id=id)
@@ -38,19 +16,24 @@ def new_item(response,id):
                         ls.item_set.create(text=txt, complete=False)
         else:
                 print("invalid")
+
         return redirect(to=f"/{id}")
 
+
 def index(response, id):
+        if response.user.is_anonymous:
+                return redirect(to=f"/login")
+
         ls = ToDoList.objects.get(id=id)
-        print("cis")
 
         if ls in response.user.todolist.all() and response.method == "POST":
                 if response.POST.get("save"):
                         list_name = response.POST.get("list_name")
+                        print(ls.name)
                         ls.name = list_name
-
+                        ls.save()
+                        print(ls.name)
                         for item in ls.item_set.all():
-                                print(item.date)
                                 if response.POST.get("c"+str(item.id)) == "clicked":
                                         item.complete = True
                                 else:
@@ -69,9 +52,7 @@ def index(response, id):
                 else:
                         Item.objects.filter(id=int(response.POST.get("remove"))).delete()
 
-
         return render(response,"main/list.html",{"ls":ls})
-        # return render(response,"main/view.html",{})
 
 def home(response):
         lists = ToDoList.objects.filter(user=response.user.id)
@@ -84,6 +65,9 @@ def home(response):
         return render(response,"main/home.html",{"are_incoming":are_incoming})
 
 def create(response):
+        if response.user.is_anonymous:
+                return redirect(to=f"/login")
+        
         if response.method == "POST":
                 form = CreateNewList(response.POST)
                 if form.is_valid():
@@ -99,6 +83,9 @@ def create(response):
       
 
 def view(response):
+        if response.user.is_anonymous:
+                return redirect(to=f"/login")
+        
         print(response.POST.get("remove"))
         if response.POST.get("remove"):
                 ToDoList.objects.filter(id=int(response.POST.get("remove"))).delete()
